+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - Moved strings to localization file.
 - Changed the scope from Global to Script in MSFT_ScheduledTask.Integration.Tests.ps1
 - Changed the scope from Global to Script ComputerManagementDsc.Common.Tests.ps1
+- ScheduledTask:
+  - Added support for event based triggers, implemented using the ScheduleType OnEvent
 
 ## 5.1.0.0