+++ CHANGELOG.md
@@ -4,6 +4,16 @@
 
 - Added Langauge Resource from LanguageDsc - fixes [Issue #127](https://github.com/PowerShell/ComputerManagementDsc/issues/127)
 - Added LangaugePack Resource from LangaugeDsc - fixes [Issue #127](https://github.com/PowerShell/ComputerManagementDsc/issues/127)
+- PowershellExecutionPolicy:
+  - Updated to meet HQRM guidelines.
+  - Migrated the xPowershellExecutionPolicy from [xPowershellExecutionPolicy](https://github.com/PowerShell/xPowerShellExecutionPolicy)
+    and renamed to PowershellExecutionPolicy.
+  - Moved strings to localization file.
+- Changed the scope from Global to Script in MSFT_ScheduledTask.Integration.Tests.ps1
+- Changed the scope from Global to Script ComputerManagementDsc.Common.Tests.ps1
+
+## 5.1.0.0
+
 - TimeZone:
   - Migrated xTimeZone resource from [xTimeZone](https://github.com/PowerShell/xTimeZone)
     and renamed to TimeZone - fixes [Issue #157](https://github.com/PowerShell/ComputerManagementDsc/issues/157).